# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WorkFlow — Work Time Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DailyTask {
  id: number;
  taskText: string;
  date: string;
  completed: boolean;
  /**
   * @minimum 1
   * @maximum 3
   */
  position: number;
  completedAt?: string | null;
  createdAt: string;
}

export interface UpsertDailyTaskBody {
  taskText: string;
  date: string;
  /**
   * @minimum 1
   * @maximum 3
   */
  position: number;
  id?: number | null;
}

export interface CompleteTaskBody {
  completed: boolean;
}

export interface TaskLogEntry {
  id: number;
  taskName: string;
  date: string;
  timeStarted?: string | null;
  timeEnded?: string | null;
  block1: boolean;
  block2: boolean;
  block3: boolean;
  block4: boolean;
  createdAt: string;
}

export interface CreateTaskLogEntryBody {
  taskName: string;
  date: string;
  timeStarted?: string | null;
  timeEnded?: string | null;
  block1: boolean;
  block2: boolean;
  block3: boolean;
  block4: boolean;
}

export interface UpdateTaskLogEntryBody {
  taskName?: string;
  timeStarted?: string | null;
  timeEnded?: string | null;
  block1?: boolean;
  block2?: boolean;
  block3?: boolean;
  block4?: boolean;
}

export interface StreakInfo {
  currentStreak: number;
  longestStreak: number;
  lastCompletedDate?: string | null;
  restoreMode: boolean;
  /** Number of tasks completed today in restore mode (need 6) */
  restoreProgress: number;
}

export interface Countdown {
  id: number;
  label: string;
  targetDate: string;
  createdAt: string;
}

export interface CreateCountdownBody {
  label: string;
  targetDate: string;
}

export interface UpdateCountdownBody {
  label?: string;
  targetDate?: string;
}

export interface StatsSummary {
  currentStreak: number;
  longestStreak: number;
  totalDaysCompleted: number;
  /** Fraction 0.0-1.0 */
  weeklyCompletionRate: number;
  /** Fraction 0.0-1.0 */
  monthlyCompletionRate: number;
  totalTaskLogEntries: number;
  restoreMode: boolean;
  restoreProgress: number;
}

export interface DayCompletion {
  date: string;
  completed: boolean;
  tasksCompleted: number;
}

export type GetDailyTasksParams = {
  date?: string;
};

export type GetTaskLogParams = {
  date?: string;
};

export type GetStatsHistoryParams = {
  year?: number;
  month?: number;
};