# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WorkFlow — Work Time Tracker API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get today's 3 non-negotiable tasks
 */
export const GetDailyTasksQueryParams = zod.object({
  date: zod.date().optional(),
});

export const getDailyTasksResponsePositionMax = 3;

export const GetDailyTasksResponseItem = zod.object({
  id: zod.number(),
  taskText: zod.string(),
  date: zod.coerce.date(),
  completed: zod.boolean(),
  position: zod.number().min(1).max(getDailyTasksResponsePositionMax),
  completedAt: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
});
export const GetDailyTasksResponse = zod.array(GetDailyTasksResponseItem);

/**
 * @summary Create or update a daily task
 */
export const upsertDailyTaskBodyPositionMax = 3;

export const UpsertDailyTaskBody = zod.object({
  taskText: zod.string(),
  date: zod.coerce.date(),
  position: zod.number().min(1).max(upsertDailyTaskBodyPositionMax),
  id: zod.number().nullish(),
});

export const upsertDailyTaskResponsePositionMax = 3;

export const UpsertDailyTaskResponse = zod.object({
  id: zod.number(),
  taskText: zod.string(),
  date: zod.coerce.date(),
  completed: zod.boolean(),
  position: zod.number().min(1).max(upsertDailyTaskResponsePositionMax),
  completedAt: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Mark a daily task as complete or incomplete
 */
export const CompleteDailyTaskParams = zod.object({
  id: zod.coerce.number(),
});

export const CompleteDailyTaskBody = zod.object({
  completed: zod.boolean(),
});

export const completeDailyTaskResponsePositionMax = 3;

export const CompleteDailyTaskResponse = zod.object({
  id: zod.number(),
  taskText: zod.string(),
  date: zod.coerce.date(),
  completed: zod.boolean(),
  position: zod.number().min(1).max(completeDailyTaskResponsePositionMax),
  completedAt: zod.coerce.date().nullish(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Get task log entries for a date
 */
export const GetTaskLogQueryParams = zod.object({
  date: zod.date().optional(),
});

export const GetTaskLogResponseItem = zod.object({
  id: zod.number(),
  taskName: zod.string(),
  date: zod.coerce.date(),
  timeStarted: zod.string().nullish(),
  timeEnded: zod.string().nullish(),
  block1: zod.boolean(),
  block2: zod.boolean(),
  block3: zod.boolean(),
  block4: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const GetTaskLogResponse = zod.array(GetTaskLogResponseItem);

/**
 * @summary Create a new task log entry
 */
export const CreateTaskLogEntryBody = zod.object({
  taskName: zod.string(),
  date: zod.coerce.date(),
  timeStarted: zod.string().nullish(),
  timeEnded: zod.string().nullish(),
  block1: zod.boolean(),
  block2: zod.boolean(),
  block3: zod.boolean(),
  block4: zod.boolean(),
});

/**
 * @summary Update a task log entry
 */
export const UpdateTaskLogEntryParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateTaskLogEntryBody = zod.object({
  taskName: zod.string().optional(),
  timeStarted: zod.string().nullish(),
  timeEnded: zod.string().nullish(),
  block1: zod.boolean().optional(),
  block2: zod.boolean().optional(),
  block3: zod.boolean().optional(),
  block4: zod.boolean().optional(),
});

export const UpdateTaskLogEntryResponse = zod.object({
  id: zod.number(),
  taskName: zod.string(),
  date: zod.coerce.date(),
  timeStarted: zod.string().nullish(),
  timeEnded: zod.string().nullish(),
  block1: zod.boolean(),
  block2: zod.boolean(),
  block3: zod.boolean(),
  block4: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a task log entry
 */
export const DeleteTaskLogEntryParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get current streak info
 */
export const GetStreaksResponse = zod.object({
  currentStreak: zod.number(),
  longestStreak: zod.number(),
  lastCompletedDate: zod.coerce.date().nullish(),
  restoreMode: zod.boolean(),
  restoreProgress: zod
    .number()
    .describe("Number of tasks completed today in restore mode (need 6)"),
});

/**
 * @summary Recalculate and update streaks after completing tasks
 */
export const UpdateStreaksResponse = zod.object({
  currentStreak: zod.number(),
  longestStreak: zod.number(),
  lastCompletedDate: zod.coerce.date().nullish(),
  restoreMode: zod.boolean(),
  restoreProgress: zod
    .number()
    .describe("Number of tasks completed today in restore mode (need 6)"),
});

/**
 * @summary Get custom countdown events
 */
export const GetCountdownsResponseItem = zod.object({
  id: zod.number(),
  label: zod.string(),
  targetDate: zod.coerce.date(),
  createdAt: zod.coerce.date(),
});
export const GetCountdownsResponse = zod.array(GetCountdownsResponseItem);

/**
 * @summary Create a countdown
 */
export const CreateCountdownBody = zod.object({
  label: zod.string(),
  targetDate: zod.coerce.date(),
});

/**
 * @summary Update a countdown
 */
export const UpdateCountdownParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateCountdownBody = zod.object({
  label: zod.string().optional(),
  targetDate: zod.coerce.date().optional(),
});

export const UpdateCountdownResponse = zod.object({
  id: zod.number(),
  label: zod.string(),
  targetDate: zod.coerce.date(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a countdown
 */
export const DeleteCountdownParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get stats summary (streak, completion rates)
 */
export const GetStatsSummaryResponse = zod.object({
  currentStreak: zod.number(),
  longestStreak: zod.number(),
  totalDaysCompleted: zod.number(),
  weeklyCompletionRate: zod.number().describe("Fraction 0.0-1.0"),
  monthlyCompletionRate: zod.number().describe("Fraction 0.0-1.0"),
  totalTaskLogEntries: zod.number(),
  restoreMode: zod.boolean(),
  restoreProgress: zod.number(),
});

/**
 * @summary Get daily completion history for calendar heatmap
 */
export const GetStatsHistoryQueryParams = zod.object({
  year: zod.coerce.number().optional(),
  month: zod.coerce.number().optional(),
});

export const GetStatsHistoryResponseItem = zod.object({
  date: zod.coerce.date(),
  completed: zod.boolean(),
  tasksCompleted: zod.number(),
});
export const GetStatsHistoryResponse = zod.array(GetStatsHistoryResponseItem);