# lib/api-zod/src/generated/types/countdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WorkFlow — Work Time Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Countdown {
  id: number;
  label: string;
  targetDate: Date;
  createdAt: Date;
}