# lib/api-zod/src/generated/types/createTaskLogEntryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WorkFlow — Work Time Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTaskLogEntryBody {
  taskName: string;
  date: Date;
  timeStarted?: string | null;
  timeEnded?: string | null;
  block1: boolean;
  block2: boolean;
  block3: boolean;
  block4: boolean;
}