# lib/api-zod/src/generated/types/dailyTask.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WorkFlow — Work Time Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyTask {
  id: number;
  taskText: string;
  date: Date;
  completed: boolean;
  /**
   * @minimum 1
   * @maximum 3
   */
  position: number;
  completedAt?: Date | null;
  createdAt: Date;
}