# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WorkFlow — Work Time Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface StatsSummary {
  currentStreak: number;
  longestStreak: number;
  totalDaysCompleted: number;
  /** Fraction 0.0-1.0 */
  weeklyCompletionRate: number;
  /** Fraction 0.0-1.0 */
  monthlyCompletionRate: number;
  totalTaskLogEntries: number;
  restoreMode: boolean;
  restoreProgress: number;
}