# lib/api-zod/src/generated/types/streakInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * WorkFlow — Work Time Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface StreakInfo {
  currentStreak: number;
  longestStreak: number;
  lastCompletedDate?: Date | null;
  restoreMode: boolean;
  /** Number of tasks completed today in restore mode (need 6) */
  restoreProgress: number;
}